60fps & Beam Error
After I made the 60 fps cap, I tested it by firing a ton of beams to see if it accounted for a ton more objects being added and drawn. It did, but if there were a LOT of beams on screen, it would throw a ConcurrentEdit error or something like that. I figured because of the fact that it's using frames now, it had to be because we were adding a new beam to the array and also getting the array, making them not match up. I fixed this by making the getBeams() return an old version of the array while the fire() method adds one to the current array. Also, I fixed a stutter due to getScaledInstance(). Apparently it's really slow and if you have a lot of tiles it makes the framerate suffer.  The normal drawImage() method can actually resize things, which I didn't know, so using that made it a lot smoother. This is a lot of words.